* vfs.c (vfs_cmp_handle): Fixing warnings on on 64-bit         platforms.         (vfs_op):Likewise.         (vfs_info):Likewise.         (vfs_free_handle):Likewise.         * extfs/rpm (mcrpmfs_run): Quote parameters to calls         to system/open that spawn a shell.         * direntry.c (vfs_s_find_entry_tree): Make "size"         size_t.

+++ vfs/direntry.c
@@ -240,7 +240,7 @@ static struct vfs_s_entry *
 vfs_s_find_entry_tree (struct vfs_class *me, struct vfs_s_inode *root,
 		       char *path, int follow, int flags)
 {
-    unsigned int pseg;
+    size_t pseg;
     struct vfs_s_entry *ent = NULL;
 
     canonicalize_pathname (path);
